[test] Refactor test_quant_api.py to use ToyLinearModel from torchao.testing.model_architectures
Partial fix for #2078.

Adds ToyLinearModel to torchao/testing/model_architectures.py as the
canonical definition, then removes the duplicate inline class from
test/quantization/test_quant_api.py and replaces it with an import.

No callsites changed — the interface is identical so this is a pure refactor.
